update_ui() and print_map()

+++ src/room.c
@@ -5,9 +5,9 @@
 static uint8_t room_h[2] = {4, 5};
 static uint8_t room_w[2] = {4, 9};
 
-static void fill_room(uint16_t y1, uint16_t x1, uint16_t y2, uint16_t x2) {
-  for (uint16_t y = y1; y <= y2; y++) {
-    for (uint16_t x = x1; x <= x2; x++) {
+static void fill_room(uint8_t y1, uint8_t x1, uint8_t y2, uint8_t x2) {
+  for (uint8_t y = y1; y <= y2; y++) {
+    for (uint8_t x = x1; x <= x2; x++) {
       m->gr[y * WIDTH + x]->gv = 2;
 
       if (m->spt >= 2) m->pt = realloc(m->pt, sizeof(m->pt) * m->spt + 1);
@@ -18,7 +18,7 @@ static void fill_room(uint16_t y1, uint16_t x1, uint16_t y2, uint16_t x2) {
 }
 
 void first_room(void) {
-  uint16_t cy, cx, h, w, y1, x1, y2, x2;
+  uint8_t cy, cx, h, w, y1, x1, y2, x2;
 
   cy = floor(HEIGHT / 2);
   cx = floor(WIDTH / 2);
